Changed matched filter pulse. Normalized fft in Sv_f and TS_f
Former-commit-id: 808a9894f27b00d6705479f207cbe1f9081ed908 [formerly 7d4eb948bb42ba2ac434a56770cdfb3e601028d1]
Former-commit-id: 71be276cca7620e49a83728b38fcc2006df66a03
Former-commit-id: bc91431da2df4b812c2e26b79408ffbcb3dc5b52
Former-commit-id: 3e9eb969c0d073db493ec6d504519bcd11ead79a [formerly d082ad217a58fd9d3c4d2cac1eff4739985ab4f0]
Former-commit-id: 5fbb170ccd112450f75122dde46a4ccda69906bf
Former-commit-id: 66c1ad0c4f7f4c6037b3625ec55198708476cee3

+++ classes/@sub_ac_data_cl/sub_ac_data_cl.m
@@ -42,7 +42,7 @@ classdef sub_ac_data_cl < handle
                         continue;
                     end
                     format={'single',size(data{icell}),field};
-                    fwrite(fileID,single(data{icell}),'single');
+                    fwrite(fileID,double(data{icell}),'single');
                     fclose(fileID);
                     
                     obj.Memap{icell} = memmapfile(curr_name,...
